requirements.txt
requests>=2.31.0
pyarrow>=14.0.0
azure-storage-blob>=12.19.0
pandas>=2.0.0
snowflake-connector-python[pandas]>=3.16.0

